% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMCAlgorithmObject.R
\name{acfMCMC}
\alias{acfMCMC}
\title{Autocorrelation function for the likelihood or posterior trace}
\usage{
acfMCMC(mcmc, type = "LogPosterior", samples = NULL, lag.max = 40, plot = TRUE)
}
\arguments{
\item{mcmc}{object of class MCMC}

\item{type}{"LogPosterior" or "LogLikelihood", defaults to "LogPosterior"}

\item{samples}{number of samples at the end of the trace used to calculate the acf}

\item{lag.max}{Maximum amount of lag to calculate acf. Default is 10*log10(N), where N i the number of observations.}

\item{plot}{logical. If TRUE (default) a plot of the acf is created}
}
\description{
The function calculates and by defaults plots the acf and estimates the autocorrelation in the trace.
}
\seealso{
\code{\link{acfCSP}}
}
