% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeObject.R
\name{addObservedSynthesisRateSet}
\alias{addObservedSynthesisRateSet}
\title{Add gene observed synthesis rates}
\usage{
addObservedSynthesisRateSet(genome, observed.expression.file,
  match.expression.by.id = TRUE)
}
\arguments{
\item{genome}{A genome object initialized with \code{\link{initializeGenomeObject}} to add observed expression data.}

\item{observed.expression.file}{A string containing the location of a file containing
empirical expression rates (optional).}

\item{match.expression.by.id}{If TRUE (default) observed expression values will be assigned by matching sequence identifier.
If FALSE observed expression values will be assigned by order}
}
\value{
gene.names Returns the names of the genes as a vector of strings.
}
\description{
\code{addObservedSynthesisRateSet} returns the observed synthesis rates of the genes within the genome specified.
}
