\name{codonToAA}
\alias{codonToAA}
\title{translates codon to amino acid}
\usage{
 codonToAA(codon)
}
\arguments{
\item{codon}{character, codon to translate}
}
\value{
Returns the amino acid encoded by the given codon as character
}
\description{
Translates a given codon into the amino acid encoded by it.
}
\seealso{
\code{\link{AAToCodon}}
}

