% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixDEta}
\alias{fixDEta}
\title{fixDEta}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Fix the value of selection its current value
}
