% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixDM}
\alias{fixDM}
\title{fixDM}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Fix the value of mutation its current value
}
