% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixSphi}
\alias{fixSphi}
\title{fixSphi}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Fix the value of s_phi (standard deviation of lognormal for synthesis rates) at its current value
}
