% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAdaptiveWidth}
\alias{getAdaptiveWidth}
\title{getAdaptiveWidth}
\value{
number of sample steps between adapting proposal widths
}
\description{
Return sample adaptiveWidth value, which is the number of samples (not iterations) between adapting parameter proposal widths
}
