% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeObject.R
\name{getCodonCountsForAA}
\alias{getCodonCountsForAA}
\title{Get Codon Counts For Each Amino Acid}
\usage{
getCodonCountsForAA(genome, aa)
}
\arguments{
\item{genome}{A genome object from which the counts of each
codon can be obtained.}

\item{aa}{A one character representation of an amino acid.}
}
\value{
Returns a matrix storing the codonCounts for the given amino acid.
}
\description{
provides the codon counts for a fiven amino acid across all genes
}
\details{
The returned matrix containes a row for each gene and a coloumn for each codon.
}
