% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCodonSpecificPosteriorMeanForCodon}
\alias{getCodonSpecificPosteriorMeanForCodon}
\title{getCodonSpecificPosteriorMeanForCodon}
\arguments{
\item{mixtureElement}{mixture to calculate CSP posterior mean. Should be between 1 and n, where n is number of mixtures.}

\item{samples}{number of samples to use for calculating posterior mean}

\item{codon}{codon to calculate CSP}

\item{paramType}{CSP to calculate posterior mean for. 0: Mutation (ROC,FONSE) or Alpha (PA, PANSE). 1: Selection (ROC,FONSE), Lambda (PANSE), Lambda^prime (PA). 2: NSERate (PANSE)}

\item{withoutReference}{If model uses reference codon, then ignore this codon (fixed at 0). Should be TRUE for ROC and FONSE. Should be FALSE for PA and PANSE.}

\item{log_scale}{If true, calculate posterior mean on log scale. Should only be used for PA and PANSE.}
}
\value{
posterior mean value for CSP
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Calculate codon-specific parameter (CSP) posterior mean
}
