% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCodonSpecificPosteriorVarianceForCodon}
\alias{getCodonSpecificPosteriorVarianceForCodon}
\title{getCodonSpecificPosteriorVarianceForCodon}
\arguments{
\item{mixtureElement}{mixture to calculate CSP variance. Should be between 1 and n, where n is number of mixtures.}

\item{samples}{number of samples to use for calculating variance}

\item{codon}{codon to calculate CSP}

\item{paramType}{CSP to calculate variance for. 0: Mutation (ROC,FONSE) or Alpha (PA, PANSE). 1: Selection (ROC,FONSE), Lambda (PANSE), Lambda^prime (PA). 2: NSERate (PANSE)}

\item{unbiased}{If TRUE, should calculate variance using unbiased (N-1). Otherwise, used biased (N) correction}

\item{withoutReference}{If model uses reference codon, then ignore this codon (fixed at 0). Should be TRUE for ROC and FONSE. Should be FALSE for PA and PANSE.}

\item{log_scale}{If true, calculate posterior mean on log scale. Should only be used for PA and PANSE.}
}
\value{
variance over trace for CSP
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Calculate codon-specific parameter (CSP) variance
}
