% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCodonSpecificQuantilesForCodon}
\alias{getCodonSpecificQuantilesForCodon}
\title{getCodonSpecificQuantilesForCodon}
\arguments{
\item{mixtureElement}{mixture to calculate CSP variance. Should be between 1 and n, where n is number of mixtures.}

\item{samples}{number of samples to use for calculating variance}

\item{codon}{codon to calculate CSP}

\item{paramType}{CSP to calculate variance for. 0: Mutation (ROC,FONSE) or Alpha (PA, PANSE). 1: Selection (ROC,FONSE), Lambda (PANSE), Lambda^prime (PA). 2: NSERate (PANSE)}

\item{probs}{vector of two doubles between 0 and 1 indicating range over which to calculate quantiles. <0.0275, 0.975> would give 95\% quantiles.}

\item{withoutReference}{If model uses reference codon, then ignore this codon (fixed at 0). Should be TRUE for ROC and FONSE. Should be FALSE for PA and PANSE.}

\item{log_scale}{If true, calculate posterior mean on log scale. Should only be used for PA and PANSE.}
}
\value{
vector representing lower and upper bound of quantile
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Calculate quantiles of CSP traces
}
