% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{getExpressionEstimatesForMixture}
\alias{getExpressionEstimatesForMixture}
\title{Returns the estimated phi posterior for a gene}
\usage{
getExpressionEstimatesForMixture(parameter, gene.index, mixtureAssignment,
  samples)
}
\arguments{
\item{parameter}{on object created by \code{initializeParameterObject}.}

\item{gene.index}{a integer or vector of integers representing the gene(s) of interesst.}

\item{mixtureAssignment}{a vector with the same length as \code{gene.index} decsribing the mixture assignment of the gene}

\item{samples}{number of samples for the posterior estimate}
}
\value{
returns a vector with the mixture assignment of each gene corresbonding to \code{gene.index} in the same order as the genome.
}
\description{
Posterior estimates for the phi value of specified genes
}
\details{
The returned vector is unnamed as gene ids are only stored in the \code{genome} object, 
but the \code{gene.index} vector can be used to match the assignment to the genome.
}
\examples{
\dontrun{
# get the estimated phi values for all genes assuming they are in mixture 1
getExpressionEstimatesForMixture(parameter, 1:length(genome), rep(1, length(genome), 1000)

# get the estimated phi values for all genes for the estimated mixture element
mix.assign <- getMixtureAssignmentEstimate(parameter, 1:length(genome), 1000)
getExpressionEstimatesForMixture(parameter, 1:length(genome), mix.assign, 1000)
}
}
