% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getGroupList}
\alias{getGroupList}
\title{getGroupList}
\value{
returns list of amino acids or codons
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Get amino acids (ROC, FONSE) or codons (PA, PANSE) for which parameters will be estimated
}
