% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLogPosteriorMean}
\alias{getLogPosteriorMean}
\title{getLogPosteriorMean}
\arguments{
\item{samples}{postive value less than total length of the MCMC trace}
}
\value{
mean logPosterior
}
\description{
Method of MCMC class (access via mcmc$<function name>, where mcmc is an object initialized by initializeMCMCObject). Calculate the mean log posterior probability over the last n samples
}
