% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLogPosteriorTrace}
\alias{getLogPosteriorTrace}
\title{getLogPosteriorTrace}
\value{
vector representing logPosterior trace
}
\description{
Method of MCMC class (access via mcmc$<function name>, where mcmc is an object initialized by initializeMCMCObject). Returns the logPosterior trace
}
