% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNoiseOffsetVariance}
\alias{getNoiseOffsetVariance}
\title{getNoiseOffsetVariance}
\arguments{
\item{index}{mixture index to use. Should be number between 0 and n-1, where n is number of mixtures}

\item{samples}{number of samples over which to calculate variance}

\item{unbiased}{If TRUE, should calculate variance using unbiased (N-1). Otherwise, used biased (N) correction}
}
\value{
returns variance for noise offset
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Calculate variance of noise offset parameter used when fitting model with empirical estimates of synthesis rates (ie. withPhi fits)
}
