% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSamples}
\alias{getSamples}
\title{getSamples}
\value{
number of samples used during MCMC
}
\description{
Method of MCMC class (access via mcmc$<function name>, where mcmc is an object initialized by initializeMCMCObject). Return number of samples set for MCMCAlgorithm object
}
