% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getStdDevSynthesisRatePosteriorMean}
\alias{getStdDevSynthesisRatePosteriorMean}
\title{getStdDevSynthesisRatePosteriorMean}
\arguments{
\item{samples}{number of samples over which to calculate posterior mean}

\item{mixture}{mixture index to use. Should be number between 0 and n-1, where n is number of mixtures}
}
\value{
returns posterior mean for standard deviation of lognormal distribution of synthesis rates
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Calculate posterior mean of standard deviation parameter of lognormal describing distribution of synthesis rates
}
