% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getStdDevSynthesisRateVariance}
\alias{getStdDevSynthesisRateVariance}
\title{getStdDevSynthesisRateVariance}
\arguments{
\item{samples}{number of samples over which to calculate variance}

\item{mixture}{mixture index to use. Should be number between 0 and n-1, where n is number of mixtures}

\item{unbiased}{If TRUE, should calculate variance using unbiased (N-1). Otherwise, used biased (N) correction}
}
\value{
returns variance for standard deviation of lognormal distribution of synthesis rates
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Calculate variance of standard deviation parameter of lognormal describing distribution of synthesis rates
}
