% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTraceObject}
\alias{getTraceObject}
\title{getTraceObject}
\value{
Trace object
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Get Trace object stored by a Parameter object. Useful for plotting certain parameter traces.
}
