% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeObject.R
\name{initializeGenomeObject}
\alias{initializeGenomeObject}
\title{Genome Initialization}
\usage{
initializeGenomeObject(
  file,
  genome = NULL,
  observed.expression.file = NULL,
  fasta = TRUE,
  positional = FALSE,
  match.expression.by.id = TRUE,
  append = FALSE
)
}
\arguments{
\item{file}{A file of coding sequences in fasta or RFPData format}

\item{genome}{A genome object can be passed in to concatenate the input file to it (optional).}

\item{observed.expression.file}{String containing the location of a file containing
empirical expression rates (optional). Default value is NULL.}

\item{fasta}{Boolean value indicating whether \code{file} argument is a
fasta file (TRUE) or an RFPData file (FALSE). Default value is TRUE.}

\item{positional}{Boolean indicating if the positional information in the RFPData file is necessary. Default value is FALSE}

\item{match.expression.by.id}{If TRUE, observed expression values will be assigned by matching sequence identifier. 
If FALSE, observed expression values will be assigned by order. Default value is TRUE.}

\item{append}{If TRUE, function will read in additional genome data to append to an existing genome.
If FALSE, genome data is cleared before reading in data (no preexisting data). Default value is FALSE.}
}
\value{
This function returns the initialized Genome object.
}
\description{
\code{initializeGenomeObject} initializes the Rcpp Genome object
}
\examples{

genome_file <- system.file("extdata", "genome.fasta", package = "AnaCoDa")
genes_file <- system.file("extdata", "more_genes.fasta", package = "AnaCoDa")
expression_file <- system.file("extdata", "expression.csv", package = "AnaCoDa")
 
## reading genome
genome <- initializeGenomeObject(file = genome_file)

## reading genome and observed expression data
genome <- initializeGenomeObject(file = genome_file, observed.expression.file = expression_file)
 
## add aditional genes to existing genome
genome <- initializeGenomeObject(file = genome_file)
genome <- initializeGenomeObject(file = genes_file, genome = genome, append = TRUE)   

}
