% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializeSynthesisRateByGenome}
\alias{initializeSynthesisRateByGenome}
\title{initializeSynthesisRateByGenome}
\arguments{
\item{genome}{a Genome object}
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Initialize synthesis rates using SCUO values calcuated from the genome
}
