% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initSelectionCategories}
\alias{initSelectionCategories}
\title{initSelectionCategories}
\arguments{
\item{files}{list of files containing starting values. Number of files should equal the number of categories.}

\item{numCategories}{number of mutation categories (should be less than or equal to number of mixtures)}

\item{fix}{Can use this parameter to fix selection at current values (won't change over course of MCMC run)}
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Initialize values for selection CSP. File should be of comma-separated with header. Three columns should be of order Amino_acid,Codon,Value
}
