% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeObject.R
\name{length.Rcpp_Genome}
\alias{length.Rcpp_Genome}
\title{Length of Genome}
\usage{
\method{length}{Rcpp_Genome}(x)
}
\arguments{
\item{x}{A genome object initialized with \code{\link{initializeGenomeObject}}.}
}
\value{
returns the number of genes in a genome
}
\description{
\code{length} gives the length of a genome
}
\examples{

genome_file <- system.file("extdata", "genome.fasta", package = "AnaCoDa")
 
## reading genome
genome <- initializeGenomeObject(file = genome_file)
length(genome) # 10

}
