% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcObject.R
\name{loadMCMCObject}
\alias{loadMCMCObject}
\title{Load MCMC Object}
\usage{
loadMCMCObject(files)
}
\arguments{
\item{files}{The filenames where the data will be stored.}
}
\value{
This function has no return value.
}
\description{
\code{loadMCMCObject} creates a new MCMC object and fills it with
the information in the file given.
}
\details{
This MCMC object is not intended to be used to do another model fitting, only
to graph the stored results.
}
\examples{

## loading mcmc objects from the filesystem
\dontrun{
# load one mcmc object
mcmc <- loadMCMCObject(files = "mcmc.Rda")

# load and combine multiple mcmc objects. Useful when using checkpointing
mcmc <- loadMCMCObject(files = c("mcmc1.Rda", "mcmc2.Rda"))
}
}
