% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMCAlgorithmObject.R
\name{plot.Rcpp_MCMCAlgorithm}
\alias{plot.Rcpp_MCMCAlgorithm}
\title{Plot MCMC algorithm}
\usage{
\method{plot}{Rcpp_MCMCAlgorithm}(x, what = "LogPosterior", zoom.window = NULL, ...)
}
\arguments{
\item{x}{An Rcpp_MCMC object initialized with \code{initializeMCMCObject}.}

\item{what}{character defining if log(Posterior) (Default) or log(Likelihood) 
options are: LogPosterior or logLikelihood}

\item{zoom.window}{A vector describing the start and end of the zoom window.}

\item{...}{Arguments to be passed to methods, such as graphical parameters.}
}
\value{
This function has no return value.
}
\description{
This function will plot the logLikelihood trace, and if the Hmisc package is installed, it will 
 plot a subplot of the logLikelihood trace with the first few samples removed.
}
