% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelObject.R
\name{plot.Rcpp_ROCModel}
\alias{plot.Rcpp_ROCModel}
\title{Plot Model Object}
\usage{
\method{plot}{Rcpp_ROCModel}(x, genome = NULL, samples = 100, mixture = 1, simulated = FALSE, ...)
}
\arguments{
\item{x}{An Rcpp model object initialized with \code{initializeModelObject}.}

\item{genome}{An Rcpp genome object initialized with \code{initializeGenomeObject}.}

\item{samples}{The number of samples in the trace}

\item{mixture}{The mixture for which to graph values.}

\item{simulated}{A boolean value that determines whether to use the simulated genome.}

\item{...}{Optional, additional arguments.
For this function, a possible title for the plot in the form of a list if set with "main".}
}
\value{
This function has no return value.
}
\description{
Plots traces from the model object such as synthesis rates for each gene.
Will work regardless of whether or not expression/synthesis rate levels are being
estimated. If you wish to plot observed/empirical values, these values MUST be set
using the initial.expression.values parameter found in initializeParameterObject.
Otherwise, the expression values plotted will just be SCUO values estimated upon
initialization of the Parameter object.
}
