% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraceObject.R
\name{plot.Rcpp_Trace}
\alias{plot.Rcpp_Trace}
\title{Plot Trace Object}
\usage{
\method{plot}{Rcpp_Trace}(x, what = c("Mutation", "Selection",
  "MixtureProbability", "Sphi", "Mphi", "Aphi", "Sepsilon", "ExpectedPhi",
  "Expression"), geneIndex = 1, mixture = 1, ...)
}
\arguments{
\item{x}{An Rcpp trace object initialized with \code{initializeTraceObject}.}

\item{what}{A string containing one of the following to graph: \code{Mutation, Selection, Alpha, LambdaPrime, MeanWaitingTime, VarWaitingTime
MixtureProbability, Sphi, Mphi, Aphi, Spesilon, ExpectedPhi, Expression}.}

\item{geneIndex}{When plotting expression, the index of the gene to be plotted.}

\item{mixture}{The mixture for which to plot values.}

\item{...}{Optional, additional arguments.
For this function, may be a logical value determining if the trace is ROC-based or not.}
}
\value{
This function has no return value.
}
\description{
Plots different traces, specified with the \code{what} parameter.
}
