% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraceObject.R
\name{plotAcceptanceRatios}
\alias{plotAcceptanceRatios}
\title{Plot Acceptance ratios}
\usage{
plotAcceptanceRatios(trace, main = "CSP Acceptance Ratio Traces")
}
\arguments{
\item{trace}{An Rcpp trace object initialized with \code{initializeTraceObject}.}

\item{main}{The title of the plot.}
}
\value{
This function has no return value.
}
\description{
Plots acceptance ratios for codon-specific parameters. Will be by amino acid for ROC and FONSE models, but will be by codon for PA and PANSE models. Note assumes estimating parameters for all codons.
}
