% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraceObject.R
\name{plotCodonSpecificHyperParameters}
\alias{plotCodonSpecificHyperParameters}
\title{Plot Codon Specific Hyper Parameter}
\usage{
plotCodonSpecificHyperParameters(trace, mixture, type = "RandomNumber",
  main = "Random Number Parameter Traces", PA = TRUE)
}
\arguments{
\item{trace}{An Rcpp trace object initialized with \code{initializeTraceObject}.}

\item{mixture}{The mixture for which to plot values.}

\item{type}{A string containing one of the following to graph: \code{Random, LoglikelihoodRatio, currLoglikelihood, propLoglikelihood, currLoglikelihoodAdjusted, propLoglikelihoodAdjusted}.}

\item{main}{The title of the plot.}

\item{PA}{A logical value determining if the Parameter was PA or not.}
}
\value{
This function has no return value.
}
\description{
Plots a codon-specific set of traces, specified with the \code{type} parameter.
}
