% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readPhiValue}
\alias{readPhiValue}
\title{readPhiValue}
\arguments{
\item{filename}{name of file to be read}
}
\description{
Method of Parameter class (access via parameter$<function name>, where parameter is an object initialized by initializeParameterObject). Read synthesis rate values from file. File should be two column file <gene_id,phi> and is expected to have a header row
}
