% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setRestartFileSettings}
\alias{setRestartFileSettings}
\title{setRestartFileSettings}
\arguments{
\item{filename}{name of restart file}

\item{interval}{number of samples (ie. iterations * thinning) between writing new restart file}

\item{multiple}{if true, will output a new restart file at each interval (file name will include sample it was written at)}
}
\description{
Method of MCMC class (access via mcmc$<function name>, where mcmc is an object initialized by initializeMCMCObject). Set restart file output name and frequency prior to running MCMC
}
