% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setSamples}
\alias{setSamples}
\title{setSamples}
\arguments{
\item{_samples}{postive value}
}
\description{
Method of MCMC class (access via mcmc$<function name>, where mcmc is an object initialized by initializeMCMCObject). Set number of samples set for MCMCAlgorithm object
}
