% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setThinning}
\alias{setThinning}
\title{setThinning}
\arguments{
\item{_thinning}{postive value}
}
\description{
Set thinning value, which is the number of iterations (total iterations = samples * thinning) not being kept
}
