% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulateGenome}
\alias{simulateGenome}
\title{simulateGenome}
\arguments{
\item{genome}{a Genome object initialized by initializeGenomeObject}
}
\description{
Method of Model class (access via model$<function name>, where model is an object initialized by initializeModelObject). Will simulate a version of the given genome using the current set of parameters stored in the Parameter object. This can be written to a FASTA file using genome$writeFasta(<filename>,simulated = TRUE).
}
