% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeObject.R
\name{summary.Rcpp_Genome}
\alias{summary.Rcpp_Genome}
\title{Summary of Genome}
\usage{
\method{summary}{Rcpp_Genome}(object, ...)
}
\arguments{
\item{object}{A genome object initialized with \code{\link{initializeGenomeObject}}.}

\item{...}{Optional, additional arguments to be passed to the main summary function 
that affect the summary produced.}
}
\value{
This function returns by default an object of class c("summaryDefault", table").
}
\description{
\code{summary} summarizes the description of a genome, such as number of genes and average gene length.
}
