% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{writeParameterObject}
\alias{writeParameterObject}
\title{Write Parameter Object to a File}
\usage{
writeParameterObject(parameter, file)
}
\arguments{
\item{parameter}{parameter on object created by \code{initializeParameterObject}.}

\item{file}{A filename that where the data will be stored.}
}
\value{
This function has no return value.
}
\description{
\code{writeParameterObject} will write the parameter object as binary to the filesystem
}
\details{
As Rcpp object are not serializable with the default R \code{save} function, 
therefore this custom save function is provided (see \link{loadParameterObject}).
}
\examples{
\dontrun{
writeParameterObject(parameter, "file.Rda")
}

}
