% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeObject.R
\name{calculateSCUO}
\alias{calculateSCUO}
\title{calculates the synonymous codon usage order (SCUO)}
\usage{
calculateSCUO(genome)
}
\arguments{
\item{genome}{A genome object initialized with \code{\link{initializeGenomeObject}}.}
}
\value{
returns the SCUO value for each gene in genome
}
\description{
\code{calculateSCUO} calulates the SCUO value for each gene in genome. Note that if a codon is absent, this will be treated as NA and will be skipped in final calculation
}
\examples{

genome_file <- system.file("extdata", "genome.fasta", package = "AnaCoDa")
 
## reading genome
genome <- initializeGenomeObject(file = genome_file)
scuo <- calculateSCUO(genome)

}
