% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{getCSPEstimates}
\alias{getCSPEstimates}
\title{Return Codon Specific Paramters (or write to csv) estimates as data.frame}
\usage{
getCSPEstimates(parameter, filename, CSP, mixture, samples)
}
\arguments{
\item{parameter}{parameter on object created by \code{initializeParameterObject}.}

\item{filename}{Posterior estimates will be written to file instead of returned if specified (format: csv).}

\item{CSP}{which type of codon specific parameter should be returned (mutation (default) or selection)}

\item{mixture}{estimates for which mixture should be returned}

\item{samples}{The number of samples used for the posterior estimates.}
}
\value{
returns a data.frame with the posterior estimates of the models 
codon specific parameters or writes it directly to a csv file if \code{filename} is specified
}
\description{
\code{getCSPEstimates} returns the codon specific
parameter estimates for a given parameter and mixture or write it to a csv file.
}
