% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{getMixtureAssignmentEstimate}
\alias{getMixtureAssignmentEstimate}
\title{Returns mixture assignment estimates for each gene}
\usage{
getMixtureAssignmentEstimate(parameter, gene.index, samples)
}
\arguments{
\item{parameter}{on object created by \code{initializeParameterObject}}

\item{gene.index}{a integer or vector of integers representing the gene(s) of interesst.}

\item{samples}{number of samples for the posterior estimate}
}
\value{
returns a vector with the mixture assignment of each gene corresbonding to \code{gene.index} in the same order as the genome.
}
\description{
Posterior estimates for the mixture assignment of specified genes
}
\details{
The returned vector is unnamed as gene ids are only stored in the \code{genome} object, 
but the \code{gene.index} vector can be used to match the assignment to the genome.
}
\examples{
\dontrun{
# get the mixture assignment for all genes
getMixtureAssignmentEstimate(parameter, 1:length(genome), 1000)

# get the mixture assignment for a subsample
getMixtureAssignmentEstimate(parameter, 5:100, 1000)
# or
getMixtureAssignmentEstimate(parameter, c(10, 30:50, 3, 90), 1000)
}
}
