% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeObject.R
\name{getObservedSynthesisRateSet}
\alias{getObservedSynthesisRateSet}
\title{Get gene observed synthesis rates}
\usage{
getObservedSynthesisRateSet(genome, simulated = FALSE)
}
\arguments{
\item{genome}{A genome object initialized with \code{\link{initializeGenomeObject}}.}

\item{simulated}{A logical value denoting if the synthesis rates to be listed are simulated or not.
The default value is FALSE.}
}
\value{
gene.names Returns the names of the genes as a vector of strings.
}
\description{
\code{getObservedSynthesisRateSet} returns the observed synthesis rates of the genes within the genome specified.
}
