% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraceObject.R
\name{plotCodonSpecificParameters}
\alias{plotCodonSpecificParameters}
\title{Plot Codon Specific Parameter}
\usage{
plotCodonSpecificParameters(
  trace,
  mixture,
  type = "Mutation",
  main = "Mutation Parameter Traces",
  ROC.or.FONSE = TRUE,
  log.10.scale = F
)
}
\arguments{
\item{trace}{An Rcpp trace object initialized with \code{initializeTraceObject}.}

\item{mixture}{The mixture for which to plot values.}

\item{type}{A string containing one of the following to graph: \code{Mutation, Selection, Alpha, LambdaPrime, MeanWaitingTime, VarWaitingTime}.}

\item{main}{The title of the plot.}

\item{ROC.or.FONSE}{A logical value determining if the Parameter was ROC/FONSE or not.}

\item{log.10.scale}{A logical value determining if figures should be plotted on the log.10.scale (default=F). Should not be applied to mutation and selection parameters estimated by ROC/FONSE.}
}
\value{
This function has no return value.
}
\description{
Plots a codon-specific set of traces, specified with the \code{type} parameter.
}
