% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcObject.R
\name{setRestartSettings}
\alias{setRestartSettings}
\title{Set Restart Settings}
\usage{
setRestartSettings(mcmc, filename, samples, write.multiple = TRUE)
}
\arguments{
\item{mcmc}{MCMC object that will run the model fitting algorithm.}

\item{filename}{Filename for the restart files to be written.}

\item{samples}{Number of samples that should occur before a file is written.}

\item{write.multiple}{Boolean that determines if multiple restart files
are written. Default value is TRUE.}
}
\value{
This function has no return value.
}
\description{
\code{setRestartSettings} sets the needed information (what the file 
is called, how often the file should be written) to write
information to restart the MCMC algorithm from a given point.
}
\details{
\code{setRestartSettings} writes a restart file every set amount of samples
that occur. Also, if write.multiple is true, instead of overwriting the previous restart
file, the sample number is prepended onto the file name and multiple rerstart files
are generated for a run.
}
