% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcObject.R
\name{writeMCMCObject}
\alias{writeMCMCObject}
\title{Write MCMC Object}
\usage{
writeMCMCObject(mcmc, file)
}
\arguments{
\item{mcmc}{MCMC object that has run the model fitting algorithm.}

\item{file}{A filename where the data will be stored.
The file should end with the extension "Rdat".}
}
\value{
This function has no return value.
}
\description{
\code{writeMCMCObject} stores the MCMC information from the 
model fitting run in a file.
}
\details{
None.
}
